backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
web3==6.11.3
openai==1.3.7
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
transformers==4.36.2
torch==2.1.1
requests==2.31.0
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.13.1
schedule==1.2.0
geopy==2.4.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
websockets==12.0
